MCMCTesting

Documentation for MCMCTesting.

MCMCTesting.ExactRankTest

MCMCTesting.ExactRankTest

MCMCTesting.TestSubject

MCMCTesting.TwoSampleGibbsTest

MCMCTesting.TwoSampleGibbsTest

MCMCTesting.TwoSampleTest

MCMCTesting.TwoSampleTest

MCMCTesting.markovchain_transition

MCMCTesting.mcmctest

MCMCTesting.sample_joint

MCMCTesting.sample_predictive

MCMCTesting.ExactRankTest — Method

ExactRankTest(n_samples, n_mcmc_steps; n_mcmc_thin)

Exact rank hypothesis testing strategy for reversible MCMC kernels. Algorithm 2 in Gandy & Scott 2021.

Arguments

n_samples::Int: Number of ranks to be simulated.

n_mcmc_steps::Int: Number of MCMC states to be simulated for simulating a single rank.

n_mcmc_thin::Int: Number of thinning applied to the MCMC chain.

Returns

pvalues: P-value computed for each dimension of the statistic returned from statistics.

Requirements

This test requires the following functions for model and kernel to be implemented:

markovchain_transition

sample_joint

Furthermore, this test explicitly assumes the following

kernel is reversible.

Applying this tests to an irreversible kernel will result in false negatives even if its stationary distribution is correct.

Keyword Arguments for Tests

When calling mcmctest or seqmcmctest, this tests has an additional keyword argument:

uniformity_test_pvalue: The p-value calculation strategy.

The default strategy is an $\chi^2$ test. Any function returning a single p-value from a uniformity hypothesis test will work. The format is as follows:

uniformity_test_pvalue(x::AbstractVector)::Real

References

MCMCTesting.TestSubject — Type

TestSubject(model, kernel)

Model and MCMC kernel obejct subject to test.

Arguments

model: Model subject to test.

kernel: MCMC kernel subject to test.

MCMCTesting.TwoSampleGibbsTest — Method

TwoSampleTest(n_samples, n_mcmc_steps; n_control, n_treatment, n_mcmc_thin)

Two-sample hypothesis testing strategy with an additional Gibbs sampling step. Modified version of Algorithm 1 in Gandy & Scott 2021 for increased power.

Arguments

n_samples::Int: Number of samples from the joint p(θ, y) used for the computing the p-values.

n_mcmc_steps::Int: Number of times the MCMC kernel is applied to initial sample from the joint. (Increasing this value improves the power of the test.)

Keyword Arguments

n_control::Int: Number of pure samples from the joint (control group). (Default: n_samples)

n_treatment::Int: Number of samples from the MCMC kernel (treatment group). (Default: n_samples)

n_mcmc_thin::Int: Number of thinning applied to the MCMC chain. The effect of this argument is the same as n_mcmc_steps.

Returns

pvalues: P-value computed for each dimension of the statistic returned from statistics.

Requirements

This test requires the following functions for model and kernel to be implemented:

markovchain_transition

sample_joint

sample_predictive

Keyword Arguments for Tests

When calling mcmctest or seqmcmctest, this tests has an additional keyword argument:

two_sample_test_pvalue: The p-value calculation strategy.

The default strategy is an approximate two-sample Kolmogorov-Smirnov test. Any function returning a p-value to two sample groups will work. The format is as follows:

two_sample_test_pvalue(x::AbstractVector, y::AbstractVector)::Real

References

MCMCTesting.TwoSampleTest — Method

TwoSampleTest(n_samples, n_mcmc_steps; n_control, n_treatment, n_mcmc_thin)

Two-sample hypothesis testing strategy. Algorithm 1 in Gandy & Scott 2021.

Arguments

n_samples::Int: Number of samples from the joint p(θ, y) used for the computing the p-values.

n_mcmc_steps::Int: Number of times the MCMC kernel is applied to initial sample from the joint. (Increasing this value improves the power of the test.)

Keyword Arguments

n_control::Int: Number of pure samples from the joint (control group). (Default: n_samples)

n_treatment::Int: Number of samples from the MCMC kernel (treatment group). (Default: n_samples)

n_mcmc_thin::Int: Number of thinning applied to the MCMC chain. The effect of this argument is the same as n_mcmc_steps.

Returns

pvalues: P-value computed for each dimension of the statistic returned from statistics.

Requirements

This test requires the following functions for model and kernel to be implemented:

markovchain_transition

sample_joint

Keyword Arguments for Tests

When calling mcmctest or seqmcmctest, this tests has an additional keyword argument:

two_sample_test_pvalue: The p-value calculation strategy.

The default strategy is an approximate two-sample Kolmogorov-Smirnov test. Any function returning a single p-value from a two-sample hypothesis test will work. The format is as follows:

two_sample_test_pvalue(x::AbstractVector, y::AbstractVector)::Real

References

MCMCTesting.markovchain_transition — Function

markovchain_transition(rng, model, kernel, θ, y)

Perform a single Markov chain transition of kernel on the previous state θ targeting the posterior of model conditioned on y.

Arguments

rng::Random.AbstractRNG: Random number generator.

model: Model forming the posterior p(θ|y) conditioned on y.

θ: Previous state of the Markov chain.

y: Data to condition on.

Returns

θ′: Next state of the Markov chain.

MCMCTesting.mcmctest — Function

mcmctest([rng,] test, subject; kwargs...)

Sample a p-value according to test for subject

Arguments

rng::Random.AbstractRNG: Random number generator. (Default: Random.default_rng().)

test::AbstractMCMCTest: Test strategy.

subject::TestSubject: MCMC algorithm and model subject to test.

Keyword Arguments

show_progress::Bool: Whether to show the progress bar. (Default: true.)

statistics: Function for computing test statistics from samples generated from the tests. (See section below for additional description.)

Check the documentation for the respective test strategy for additional keyword arugments.

Custom Test Statistics

The statistics used for the hypothesis tests can be modified by passing a custom funciton to statistics. The default statistics are the first and second moments computed as below.

statistics = params -> vcat(params, params.^2)

The cross-interaction can also be tested by adding an additional entry as below.

statistics = params -> vcat(params, params.^2, reshape(params*params',:))

But naturally, adding more statistics increase the computational cost of computing the tests.

Also, different tests may result in different statistics being computed through the same statistics function. For example, the two-sample test strategies generate both model parameters θ and data y. Therefore, params = vcat(θ, y). On the other hand, the exac rank test only generates model parameters θ. Therefore, params = θ. Naturally, statistics can also be used to select a subset of parameters used for testing. For example, for the two-sample test strategies, if we only want to use θ for the tests, where d = length(θ) > 0, one can do the following:

statistics = params -> θ[1:d]

MCMCTesting.sample_joint — Function

sample_joint(rng, model)

Sample from the joint distribution of the prior and the predictive distribution of model.

Arguments

rng::Random.AbstractRNG: Random number generator.

model: Model subject to test.

Returns

θ: Model parameter sampled from the prior p(θ).

y: Data generated from conditionally on θ from p(y|θ)

MCMCTesting.sample_predictive — Function

sample_predictive(rng, model, θ)

Sample from the predictive distribution of model conditionally on θ

Arguments

rng::Random.AbstractRNG: Random number generator.

model: Model subject to test.

θ: Model parameters to condition on.

Returns

y: Data generated from conditionally on θ from p(y|θ)